Updating the french versions of lists
I think I made an error in the previous update to the Papilionoideae and Detarioideae pages, so I restored the previous updated versions (with corrections in author names), by copy and pasting from history. I searched and replaced [WCVP] with [POWO] and wcvp.science.kew.org/taxon/ with powo.science.kew.org/taxon/urn:lsid:ipni.org:names: in these 2 lists and in the species list.

I corrected Duparquetia Baill to Duparquetia Baill. (in French and English versions of genus page and species list page).

I updated the French versions of lists by copy-pasting the English versions of the lists and search and replacing /taxonomy/ with /fr/taxonomy/ to update with links. I did this for all pages. For the species list, I also replaced (/taxonomy/ with (/fr/taxonomy/.